Rename python source
