+++ legacyman_parser/parse_class_attributes_from_tonals.py
