+++ legacyman_parser/parser.py
@@ -7,7 +7,7 @@ from crawler.simple_crawler import SimpleCrawler
 from legacy_publisher import json_publisher
 from legacy_tester.parsed_json_tester import parsed_json_tester
 from legacyman_parser.parse_abbreviations import parse_abbreviations, ABBREVIATIONS
-from legacyman_parser.parse_class_attributes_from_tonals_colspan_insensitive_way import extract_class_attributes_from_tonals_page
+from legacyman_parser.parse_class_attributes_from_tonals import extract_class_attributes_from_tonals_page
 from legacyman_parser.parse_countries import extract_countries_in_region, COUNTRY_COLLECTION, COUNTRY_TABLE_NOT_FOUND
 from legacyman_parser.parse_flag_of_country import extract_flag_of_country, COUNTRY_FLAG_COLLECTION, \
     extract_flag_of_ns_country
